add remove jobs
