+++ README.md
@@ -7,8 +7,9 @@
 ```
 docker pull quay.io/opsnow-tools/elasticsearch-snapshot
 
-export AWS_BUCKET=elasticsearch-snapshot
 export AWS_REGION=ap-northeast-2
+export AWS_BUCKET=elasticsearch-snapshot
 export ES_HOST=http://elasticsearch.domain.com/
 export SLACK_TOKEN=xoxb-0000-0000-xxxx
+export SLACK_CHANNAL=#sandbox
 ```
